add autorizações em endpoint de userController -add que apenas admin e o user com o id logado pode acessa endpoints -add que apenas admin pode acessar endpoint de findAllUsers -fazendo metodo helper que verifica se o user esta logado ou tem admin como role

+++ backend/src/main/java/com/rian/task_manager/domain/user/UserService.java
@@ -1,14 +1,14 @@
 package com.rian.task_manager.domain.user;
 
 
-import com.rian.task_manager.exceptions.EmailAlredyExistsException;
+import com.rian.task_manager.domain.auth.SecurityService;
+import org.springframework.security.access.AccessDeniedException;
 import com.rian.task_manager.domain.task.dto.TaskResponse;
 import com.rian.task_manager.domain.user.dto.UserRequest;
 import com.rian.task_manager.domain.user.dto.UserResponse;
 import com.rian.task_manager.exceptions.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -17,12 +17,15 @@ import java.util.Optional;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final SecurityService securityService;
 
     public void salvar(User user) {
         userRepository.save(user);
     }
 
-    public UserResponse findById(Long id){
+    public UserResponse findById(Long id) {
+        //compara o id do user com o id autenticado
+        checkAccessUser(id);
         User user = findUserById(id);
         return UserResponse.fromEntity(user);
     }
@@ -33,6 +36,7 @@ public class UserService {
     }
 
     public List<TaskResponse> findAllTasksByUser(Long id){
+        checkAccessUser(id);
         User user = findUserById(id);
         return user.getTasks().stream()
                 .map(task -> TaskResponse.fromEntity(task))
@@ -40,26 +44,20 @@ public class UserService {
     }
 
     public Optional<User> findByEmail(String userEmail){
-        return userRepository.findByEmail(userEmail);
+        Optional<User> user = userRepository.findByEmail(userEmail);
+        return user;
     }
 
-    public UserResponse create(UserRequest userRequest){
-        if(userRepository.existsByEmail(userRequest.email())){
-            throw new EmailAlredyExistsException("Email já cadastrado");
-        }
-        User user = userRequest.toEntity();
-        userRepository.save(user);
-        return UserResponse.fromEntity(user);
 
-    }
-
-    public void deleteById(Long id){
-        User user = findUserById(id);
+    public void deleteById(Long id) {
+        checkAccessUser(id);
+        findUserById(id);
         userRepository.deleteById(id);
     }
 
 
     public UserResponse atualizar(Long id,UserRequest usuarioRequest) {
+        checkAccessUser(id);
         User user = findUserById(id);
         user.setName(usuarioRequest.name());
         user.setEmail(usuarioRequest.email());
@@ -68,10 +66,23 @@ public class UserService {
         return UserResponse.fromEntity(user);
     }
 
-    public User findUserById(Long id){
+
+    //metodos/funções helpers
+    private User findUserById(Long id){
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("user não encontrada!"));
         return user;
+    }
+
+    public void checkAccessUser(Long userId) {
+        User userLogado = securityService.getAuthenticatedUser();
+
+        boolean isAdmin = userLogado.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin && !userLogado.getId().equals(userId)) {
+            throw new AccessDeniedException("Acesso negado.");
+        }
     }
 
 }
